Puzzle Place puzzle files
I created these files for Puzzle Place. One is the logo which is also a puzzle, the others are puzzles I plan to cut out and gift at unSCruz and puzzles in the Puzzle Place collection that are missing pieces. The 9 hats puzzle is my own design based on the newly discovered aperiodic monotile.